rename to StatusProto
